Move ScopedEnvVar to new header file, protect discover() call
